backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic[email]==2.6.4
supabase==2.4.5
python-dotenv==1.0.1
requests==2.31.0
twilio==8.2.1
qrcode[pil]==7.4.2
PyJWT==2.8.0
python-multipart==0.0.9
